AssistantDispatcher: Add platform data to the converse() API
